requirements.txt
django==5.0.6
djangorestframework-jwt==1.11.0
django-rest-framework==0.1.0
djangorestframework-simplejwt==5.3.1